+++ lua/rustaceanvim/commands/parent_module.lua
@@ -1,5 +1,8 @@
 local M = {}
 
+local rl = require('rustaceanvim.rust_analyzer')
+local compat = require('rustaceanvim.compat')
+
 local function get_params()
   return vim.lsp.util.make_position_params(0, nil)
 end
@@ -12,7 +15,7 @@ local function handler(_, result, ctx)
 
   local location = result
 
-  if vim.tbl_islist(result) then
+  if compat.islist(result) then
     location = result[1]
   end
 
@@ -21,8 +24,6 @@ local function handler(_, result, ctx)
     vim.lsp.util.jump_to_location(location, client.offset_encoding)
   end
 end
-
-local rl = require('rustaceanvim.rust_analyzer')
 
 --- Sends the request to rust-analyzer to get the parent modules location and open it
 function M.parent_module()