+++ lua/rustaceanvim/neotest/init.lua
@@ -27,11 +27,13 @@
 
 ---@diagnostic disable: duplicate-set-field
 
+local cargo = require('rustaceanvim.cargo')
+local compat = require('rustaceanvim.compat')
+local config = require('rustaceanvim.config.internal')
 local lib = require('neotest.lib')
 local nio = require('nio')
-local trans = require('rustaceanvim.neotest.trans')
-local cargo = require('rustaceanvim.cargo')
 local overrides = require('rustaceanvim.overrides')
+local trans = require('rustaceanvim.neotest.trans')
 
 ---@package
 ---@type neotest.Adapter
@@ -41,7 +43,7 @@ local NeotestAdapter = { name = 'rustaceanvim' }
 ---@param file_name string
 ---@return string | nil
 NeotestAdapter.root = function(file_name)
-  return cargo.get_root_dir(file_name)
+  return cargo.get_config_root_dir(config.server, file_name)
 end
 
 ---@package
@@ -62,23 +64,53 @@ end
 ---@class rustaceanvim.neotest.Position: neotest.Position
 ---@field runnable? RARunnable
 
+----@param name string
+----@return integer
+local function find_buffer_by_name(name)
+  for _, bufnr in ipairs(vim.api.nvim_list_bufs()) do
+    local buf_name = vim.api.nvim_buf_get_name(bufnr)
+    if buf_name == name then
+      return bufnr
+    end
+  end
+  return 0
+end
+
+---@package
+---@class nio.rustaceanvim.Client: nio.lsp.Client
+---@field request nio.rustaceanvim.RequestClient Interface to all requests that can be sent by the client
+---@field config vim.lsp.ClientConfig
+
+---@package
+---@class nio.rustaceanvim.RequestClient: nio.lsp.RequestClient
+---@field experimental_runnables fun(args: nio.lsp.types.ImplementationParams, bufnr: integer?, opts: nio.lsp.RequestOpts): nio.lsp.types.ResponseError|nil, RARunnable[]|nil
+
 ---@package
 ---@param file_path string
 ---@return neotest.Tree
 NeotestAdapter.discover_positions = function(file_path)
   ---@type rustaceanvim.neotest.Position[]
   local positions = {}
-  local rust_analyzer = require('rustaceanvim.rust_analyzer')
-  local future = nio.control.future()
-  rust_analyzer.file_request(file_path, 'experimental/runnables', nil, function(err, runnables)
-    if err then
-      future.set_error(err)
-    else
-      future.set(runnables)
-    end
-  end)
-  local ok, runnables = pcall(future.wait)
-  if not ok or type(runnables) ~= 'table' or #runnables == 0 then
+
+  local lsp_client = require('rustaceanvim.rust_analyzer').get_client_for_file(file_path, 'experimental/runnables')
+  if not lsp_client then
+    ---@diagnostic disable-next-line: missing-parameter
+    return lib.positions.parse_tree(positions)
+  end
+  local nio_client = nio.lsp.get_client_by_id(lsp_client.id)
+  ---@cast nio_client nio.rustaceanvim.Client
+  local bufnr = find_buffer_by_name(file_path)
+  local params = {
+    textDocument = {
+      uri = vim.uri_from_fname(file_path),
+    },
+    position = nil,
+  }
+  local err, runnables = nio_client.request.experimental_runnables(params, bufnr, {
+    timeout = 100000,
+  })
+
+  if err or type(runnables) ~= 'table' or #runnables == 0 then
     ---@diagnostic disable-next-line: missing-parameter
     return lib.positions.parse_tree(positions)
   end
@@ -165,7 +197,7 @@ NeotestAdapter.discover_positions = function(file_path)
   -- sort positions by their start range
   local function sort_positions(to_sort)
     for _, item in ipairs(to_sort) do
-      if vim.tbl_islist(item) then
+      if compat.islist(item) then
         sort_positions(item)
       end
     end
@@ -173,8 +205,8 @@ NeotestAdapter.discover_positions = function(file_path)
     -- pop header from the list before sorting since it's used to sort in its parent's context
     local header = table.remove(to_sort, 1)
     table.sort(to_sort, function(a, b)
-      local a_item = vim.tbl_islist(a) and a[1] or a
-      local b_item = vim.tbl_islist(b) and b[1] or b
+      local a_item = compat.islist(a) and a[1] or a
+      local b_item = compat.islist(b) and b[1] or b
       if a_item.range[1] == b_item.range[1] then
         return a_item.name < b_item.name
       else