+++ lua/rustaceanvim/rust_analyzer.lua
@@ -54,56 +54,14 @@ M.buf_request = function(bufnr, method, params, handler)
   return client_found
 end
 
-----@param name string
-----@return integer
-local function find_buffer_by_name(name)
-  for _, bufnr in ipairs(vim.api.nvim_list_bufs()) do
-    local buf_name = vim.api.nvim_buf_get_name(bufnr)
-    if buf_name == name then
-      return bufnr
-    end
-  end
-  return 0
-end
-
 ---@param file_path string Search for clients with a root_dir matching this file path
 ---@param method string LSP method name
----@param params table|nil Parameters to send to the server
----@param handler? lsp.Handler See |lsp-handler|
----       If nil, follows resolution strategy defined in |lsp-handler-configuration|
-M.file_request = function(file_path, method, params, handler)
-  local client_found = false
+---@return lsp.Client|nil
+M.get_client_for_file = function(file_path, method)
   for _, client in ipairs(M.get_active_rustaceanvim_clients(nil, { method = method })) do
     local root_dir = client.config.root_dir
     if root_dir and vim.startswith(os.normalize_path_on_windows(file_path), root_dir) then
-      local bufnr = find_buffer_by_name(file_path)
-      if not params then
-        params = {
-          textDocument = {
-            uri = vim.uri_from_fname(file_path),
-          },
-          position = nil,
-        }
-      end
-      client.request(method, params, handler, bufnr)
-      client_found = true
-      if bufnr == -1 then
-        return
-      end
-    end
-  end
-  if not client_found then
-    local error_msg = 'No rust-analyzer client for ' .. method .. ' and file ' .. file_path
-    if handler then
-      ---@type lsp.HandlerContext
-      local ctx = {
-        bufnr = -1,
-        client_id = -1,
-        method = method,
-      }
-      handler({ code = -1, message = error_msg }, nil, ctx)
-    else
-      vim.notify(error_msg, vim.log.levels.ERROR)
+      return client
     end
   end
 end